refactor(boundaries): eliminate legacy src/lib folder
