+++ src/modules/auth/AuthProvider.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { MemoryRouter } from 'react-router-dom'
 import { describe, expect, it, vi } from 'vitest'
-import { buildApiUrl } from '@lib/api'
+import { buildApiUrl } from '@shared/lib'
 import { AuthProvider } from './AuthProvider'
 import { useAuth } from './authContext'
 import { AUTH_SESSION_STORAGE_KEY } from './storage'
