+++ src/modules/auth/LoginPage.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import { MemoryRouter, Route, Routes } from 'react-router-dom'
 import { describe, expect, it, vi } from 'vitest'
 import { AUTH_SESSION_STORAGE_KEY, AuthProvider } from '@auth'
-import { buildApiUrl } from '@lib/api'
+import { buildApiUrl } from '@shared/lib'
 import { LoginPage } from './LoginPage'
 
 type RenderLoginPageOptions = {
