+++ src/modules/auth/RegisterPage.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import { MemoryRouter, Route, Routes } from 'react-router-dom'
 import { describe, expect, it, vi } from 'vitest'
 import { AuthProvider } from '@auth'
-import { buildApiUrl } from '@lib/api'
+import { buildApiUrl } from '@shared/lib'
 import { LoginPage } from './LoginPage'
 import { RegisterPage } from './RegisterPage'
