+++ src/modules/scene-editor/CreateScenePage.tags.test.tsx
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import userEvent from '@testing-library/user-event'
-import { buildApiUrl } from '@lib/api'
+import { buildApiUrl } from '@shared/lib'
 import {
   addTagFromSearch,
   mockCreateScenePageFetch,
