+++ src/modules/scene-editor/SceneEditorShell.tsx
@@ -17,7 +17,7 @@ import {
   SKYBOX_OPTIONS,
   toDegrees,
   toRadians,
-} from "@lib/sceneEditor";
+} from "./sceneEditor";
 import {
   EDITOR_SECTIONS,
   PLAYLIST_OPTIONS,
