+++ src/modules/scene-editor/embeddedShaderScenes.ts
