+++ src/modules/scene-editor/sceneThumbnailUpload.ts
