+++ src/modules/scene-editor/test-fixtures.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import { MemoryRouter, Route, Routes } from 'react-router-dom'
 import { vi } from 'vitest'
 import { AuthProvider, type AuthenticatedUser } from '@auth'
-import { buildApiUrl } from '@lib/api'
+import { buildApiUrl } from '@shared/lib'
 import { buildAuthenticatedUser, storeAuthenticatedSession } from '@shared/test/auth'
 import { jsonResponse } from '@shared/test/http'
 import { CreateScenePage } from './CreateScenePage'
