+++ tsconfig.app.json
@@ -10,7 +10,6 @@
       "@app/*": ["./src/app/*"],
       "@auth": ["./src/modules/auth/index.ts"],
       "@auth/*": ["./src/modules/auth/*"],
-      "@lib/*": ["./src/lib/*"],
       "@modules/*": ["./src/modules/*"],
       "@shared/*": ["./src/shared/*"],
       "@test/*": ["./src/test/*"],
